move translation to own class and implement imports of concepts files
